Implement permissions for version 1.1.0

+++ src/main/java/net/xyfe/spigot/jawp/command/CommandDelWarp.java
@@ -14,6 +14,7 @@ import net.xyfe.spigot.jawp.JAWP;
  */
 public class CommandDelWarp implements CommandExecutor, TabCompleter {
   public static final String name = "delwarp";
+  public static final String permission = "jawp.command.delwarp";
 
   @Override
   public boolean onCommand(CommandSender sender, Command cmd, String label, String[] args) {
@@ -23,6 +24,10 @@ public class CommandDelWarp implements CommandExecutor, TabCompleter {
       sender.sendMessage("There must be at least 1 argument.");
       return true;
     } else {
+      if (!CmdUtil.hasPermission(sender, permission)) {
+        return true;
+      }
+
       boolean success = JAWP.instance.manager.deleteWarp(args[0]);
       if (!success) {
         sender.sendMessage("Warp doesn't exist.");
